backend/requirements.txt
-i https://pypi.org/simple
alembic==1.13.0; python_version >= '3.8'
apispec[marshmallow]==6.3.0; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.2.0; python_version < '3.11'
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-migrate==4.0.5
flask-smorest==0.42.3
flask-sqlalchemy==3.1.1
greenlet==3.0.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pytest==7.4.3
python-dotenv==1.0.0
sqlalchemy==2.0.23; python_version >= '3.7'
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.8.0; python_version >= '3.8'
webargs==8.3.0; python_full_version >= '3.7.2'
werkzeug==3.0.1; python_version >= '3.8'
